requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
face-recognition==1.3.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
deepface==0.0.79
tensorflow==2.14.0
pillow==10.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
matplotlib==3.8.2
seaborn==0.13.0 